+++ src/model/SharpenImage.java
@@ -8,7 +8,7 @@ public class SharpenImage extends ImageData implements ISharpen{
 
   @Override
   public int[][][] applySharpeningEffect() {
-    double[][] sharpen = {{-1.0f/8.0f, -1.0f/8.0f, -1.0f/8.0f, -1.0f/8.0f, -1.0f/8.0f},
+    float[][] sharpen = {{-1.0f/8.0f, -1.0f/8.0f, -1.0f/8.0f, -1.0f/8.0f, -1.0f/8.0f},
             {-1.0f/8.0f, 1.0f/4.0f, 1.0f/4.0f, 1.0f/4.0f, -1.0f/8.0f},
             {-1.0f/8.0f, 1.0f/4.0f, 1.0f, 1.0f/4.0f, -1.0f/8.0f},
             {-1.0f/8.0f, 1.0f/4.0f, 1.0f/4.0f, 1.0f/4.0f, -1.0f/8.0f},
